requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0
backports.zoneinfo==0.2.1; python_version < '3.9'
dj-database-url==0.5.0
django-cors-headers==3.11.0
django-environ==0.8.1
django-heroku==0.3.1
django==4.0.2
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
pytz==2021.3
setuptools==60.9.3; python_version >= '3.7'
sqlparse==0.4.2
whitenoise==6.0.0
